+++ vm/src/vm/runners/cairo_runner.rs
@@ -7,6 +7,7 @@ use crate::{
         collections::{BTreeMap, HashMap, HashSet},
         ops::{Add, AddAssign, Mul, MulAssign, Sub, SubAssign},
         prelude::*,
+        rc::Rc,
     },
     types::{builtin_name::BuiltinName, layout::CairoLayoutParams, layout_name::LayoutName},
     vm::{
@@ -643,6 +644,8 @@ impl CairoRunner {
         references: &[HintReference],
         hint_executor: &mut dyn HintProcessor,
     ) -> Result<Vec<Box<dyn Any>>, VirtualMachineError> {
+        let constants = Rc::new(self.program.constants.clone());
+
         self.program
             .shared_program_data
             .hints_collection
@@ -654,6 +657,7 @@ impl CairoRunner {
                         &hint.flow_tracking_data.ap_tracking,
                         &hint.flow_tracking_data.reference_ids,
                         references,
+                        constants.clone(),
                     )
                     .map_err(|_| VirtualMachineError::CompileHintFail(hint.code.clone().into()))
             })
@@ -704,6 +708,7 @@ impl CairoRunner {
                     .unwrap_or(&[]),
                 #[cfg(feature = "extensive_hints")]
                 &mut hint_ranges,
+                #[cfg(feature = "test_utils")]
                 &self.program.constants,
             )?;
 
@@ -760,6 +765,7 @@ impl CairoRunner {
                 hint_data,
                 #[cfg(feature = "extensive_hints")]
                 &mut hint_ranges,
+                #[cfg(feature = "test_utils")]
                 &self.program.constants,
             )?;
         }